+++ Changelog.md
@@ -2,6 +2,8 @@
 * 2 minor enhancements
   * updated document.from_library? - now uses source_type
   * added document.is_a_i?
+* 1 major enhancement
+  * attributes defined as boolean will always return boolean values
 
 ### 2.0.5
 * 1 minor enhancement